# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Technician Hub API",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/technician-forms/work-report/select": {
      "post": {
        "tags": [
          "work-reports"
        ],
        "summary": "Select Job",
        "operationId": "select_job_api_technician_forms_work_report_select_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectJob"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/technician-forms/work-report/submit": {
      "post": {
        "tags": [
          "work-reports"
        ],
        "summary": "Submit",
        "operationId": "submit_api_technician_forms_work_report_submit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/technicians/{technician_id}/work-reports": {
      "get": {
        "tags": [
          "work-reports"
        ],
        "summary": "Reports",
        "operationId": "reports_api_technicians__technician_id__work_reports_get",
        "parameters": [
          {
            "name": "technician_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Technician Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/work-reports/{report_id}": {
      "get": {
        "tags": [
          "work-reports"
        ],
        "summary": "Report",
        "operationId": "report_api_work_reports__report_id__get",
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Report Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/technicians/{technician_id}/work-report-sessions/{session_id}/revoke": {
      "post": {
        "tags": [
          "work-reports"
        ],
        "summary": "Revoke",
        "operationId": "revoke_api_technicians__technician_id__work_report_sessions__session_id__revoke_post",
        "parameters": [
          {
            "name": "technician_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Technician Id"
            }
          },
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FormRead": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "OPEN",
              "SUBMITTED"
            ],
            "title": "Status"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "jobs": {
            "items": {
              "$ref": "#/components/schemas/JobRead"
            },
            "type": "array",
            "title": "Jobs"
          },
          "selected": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "report_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Report Id"
          },
          "payment_choices": {
            "items": {
              "type": "string",
              "enum": [
                "CASH",
                "ZELLE",
                "CHECK",
                "CREDIT_CARD",
                "VENMO",
                "SUPER",
                "ESTIMATE",
                "CANCEL"
              ]
            },
            "type": "array",
            "title": "Payment Choices",
            "default": [
              "CASH",
              "ZELLE",
              "CHECK",
              "CREDIT_CARD",
              "VENMO",
              "SUPER",
              "ESTIMATE",
              "CANCEL"
            ]
          },
          "zero_amount_choices": {
            "items": {
              "type": "string",
              "enum": [
                "CASH",
                "ZELLE",
                "CHECK",
                "CREDIT_CARD",
                "VENMO",
                "SUPER",
                "ESTIMATE",
                "CANCEL"
              ]
            },
            "type": "array",
            "title": "Zero Amount Choices"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status",
          "expires_at"
        ],
        "title": "FormRead"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobRead": {
        "properties": {
          "choice_id": {
            "type": "string",
            "format": "uuid",
            "title": "Choice Id"
          },
          "operational_date": {
            "type": "string",
            "format": "date",
            "title": "Operational Date"
          },
          "start_time": {
            "type": "string",
            "title": "Start Time"
          },
          "end_time": {
            "type": "string",
            "title": "End Time"
          },
          "sequence": {
            "type": "integer",
            "title": "Sequence"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "submitted": {
            "type": "boolean",
            "title": "Submitted",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "choice_id",
          "operational_date",
          "start_time",
          "end_time",
          "sequence",
          "title",
          "location"
        ],
        "title": "JobRead"
      },
...
      "ReportInput": {
        "properties": {
          "amount_closed": {
            "type": "string",
            "title": "Amount Closed"
          },
          "payment_method": {
            "type": "string",
            "enum": [
              "CASH",
              "ZELLE",
              "CHECK",
              "CREDIT_CARD",
              "VENMO",
              "SUPER",
              "ESTIMATE",
              "CANCEL"
            ],
            "title": "Payment Method"
          },
          "closed_by": {
            "type": "string",
            "enum": [
              "MYSELF",
              "CALL_CENTER"
            ],
            "title": "Closed By"
          },
          "comments": {
            "type": "string",
            "maxLength": 4000,
            "title": "Comments",
            "default": ""
          },
          "yearly_maintenance_plan_provided": {
            "type": "boolean",
            "title": "Yearly Maintenance Plan Provided"
          },
          "reviews": {
            "$ref": "#/components/schemas/Reviews"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "amount_closed",
          "payment_method",
          "closed_by",
          "yearly_maintenance_plan_provided",
          "reviews"
        ],
        "title": "ReportInput"
      },
      "ReportList": {
        "properties": {
          "reports": {
            "items": {
              "$ref": "#/components/schemas/ReportRead"
            },
            "type": "array",
            "title": "Reports"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "reports",
          "limit"
        ],
        "title": "ReportList"
      },
      "ReportRead": {
        "properties": {
          "amount_closed": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,10}|(?=[\\d.]{1,13}0*$)\\d{0,10}\\.\\d{0,2}0*$)",
            "title": "Amount Closed"
          },
          "payment_method": {
            "type": "string",
            "enum": [
              "CASH",
              "ZELLE",
              "CHECK",
              "CREDIT_CARD",
              "VENMO",
              "SUPER",
              "ESTIMATE",
              "CANCEL"
            ],
            "title": "Payment Method"
          },
          "closed_by": {
            "type": "string",
            "enum": [
              "MYSELF",
              "CALL_CENTER"
            ],
            "title": "Closed By"
          },
          "comments": {
            "type": "string",
            "maxLength": 4000,
            "title": "Comments",
            "default": ""
          },
          "yearly_maintenance_plan_provided": {
            "type": "boolean",
            "title": "Yearly Maintenance Plan Provided"
          },
          "reviews": {
            "$ref": "#/components/schemas/Reviews"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "technician_id": {
            "type": "string",
            "format": "uuid",
            "title": "Technician Id"
          },
          "technician_name": {
            "type": "string",
            "title": "Technician Name"
          },
          "operational_date": {
            "type": "string",
            "format": "date",
            "title": "Operational Date"
          },
          "start_time": {
            "type": "string",
            "title": "Start Time"
          },
          "end_time": {
            "type": "string",
            "title": "End Time"
          },
          "sequence": {
            "type": "integer",
            "title": "Sequence"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "revision_number": {
            "type": "integer",
            "title": "Revision Number"
          },
          "submitted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Submitted At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "amount_closed",
          "payment_method",
          "closed_by",
          "yearly_maintenance_plan_provided",
          "reviews",
          "id",
          "technician_id",
          "technician_name",
          "operational_date",
          "start_time",
          "end_time",
          "sequence",
          "title",
          "location",
          "revision_number",
          "submitted_at"
        ],
        "title": "ReportRead"
      },
...
      "Reviews": {
        "properties": {
          "GOOGLE": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Google"
          },
          "GROUPON": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Groupon"
          },
          "FACEBOOK": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Facebook"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "GOOGLE",
          "GROUPON",
          "FACEBOOK"
        ],
        "title": "Reviews"
      },
...
      "SelectJob": {
        "properties": {
          "choice_id": {
            "type": "string",
            "format": "uuid",
            "title": "Choice Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "choice_id"
        ],
        "title": "SelectJob"
      },
...
      "SubmissionRead": {
        "properties": {
          "report_id": {
            "type": "string",
            "format": "uuid",
            "title": "Report Id"
          },
          "revision_number": {
            "type": "integer",
            "title": "Revision Number",
            "default": 1
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Report submitted successfully."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "report_id"
        ],
        "title": "SubmissionRead"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}